+++ challenge.py
@@ -1,0 +1,8 @@
+whole_word =
+A = 
+B =
+if len(whole_word) %2 !=0:
+    return -1
+
+def logical_comparison:
+
